test: migrate searchAction tests in source
